+++ app/javascript/components/record-form/form/subforms/subform-drawer/component.jsx
@@ -12,8 +12,9 @@ import css from "./styles.css";
 
 const Component = ({ open, cancelHandler, children, title }) => {
   return (
-    <Drawer data-testid="drawer" anchor="right" open={open} onClose={cancelHandler} classes={{ paper: css.subformDrawer }}>
-      <div className={css.subformDrawerContent}>
+
+    <Drawer anchor="right" open={open} onClose={cancelHandler} classes={{ paper: css.subformDrawer }}>
+      <div data-testid="subForm-drawer" className={css.subformDrawerContent}>
         <div className={css.title}>
           <h1>{title}</h1>
           <div>
